Struct bitmap32::slice::BitSlice

#[repr(transparent)]
pub struct BitSlice<B: BitBlock>(pub [B]);

Represents a slice of bits. B is the internal representation of the blocks used for the BitSlice

Tuple Fields
0: [B]
Implementations

impl<B: BitBlock> BitSlice<B>

pub fn len_blocks(&self) -> usize

Returns the length of self in blocks.

pub fn len_bits(&self) -> usize

Returns the length of self in bits.

pub fn get_blocks<I: SliceIndex<[B]>>(&self, index: I) -> Option<&I::Output>

pub fn get_blocks_mut<I: SliceIndex<[B]>>( &mut self, index: I ) -> Option<&mut I::Output>

pub fn blocks(&self) -> Iter<'_, B>

Iterates over the blocks in Self

pub fn blocks_mut(&mut self) -> IterMut<'_, B>

Mutable iterates over the blocks in Self

pub fn bits(&self) -> Bits<'_, B>

Iterates over the bits in self.

pub fn ones(&self) -> Ones<'_, B>

Iterates over the indices of the 1s in self. This is equivalent to self.bits().enumerate().filter_map(|(i, bit)| bit.then_some(i)), but it makes use of several optimizations and usually a count leading zero instruction or something similar.

pub fn zeros(&self) -> Zeros<'_, B>

Iterates over the indices of the 0s in self. This is equivalent to self.bits().enumerate().filter_map(|(i, bit)| (!bit).then_some(i)), but it makes use of several optimizations and usually a count leading zero instruction or something similar.

pub fn test(&self, idx: usize) -> Option<bool>

Returns true if the bit at idx is a 1. If idx is out of range, None is returned.

pub fn set_checked(&mut self, idx: usize) -> bool

Sets the bit at idx to 1. If idx is out of range, false is returned.

pub fn reset_checked(&mut self, idx: usize) -> bool

Sets the bit at idx to 0. If idx is out of range, false is returned.

pub fn clear(&mut self)

Zeros all of the bits in self.

pub fn bit_or_assign_iter<I: IntoIterator<Item = B>>(&mut self, other: I)

Does a BitOrAssign between rhs and self. If rhs is bigger than self, the overflowing bits are ignored.

pub fn bit_or_assign(&mut self, other: &Self)

Does a BitOrAssign between rhs and self. If rhs is bigger than self, the overflowing bits are ignored.
This function is purely a convinience function for Self::bit_or_assign_iter.

Note: the reason why the BitOrAssign trait isn’t used is because it requires that Self is Sized.

pub fn fill_ones_checked(&mut self, start: usize, length: NonZeroUsize) -> bool

pub fn fill_zeros_checked(&mut self, start: usize, length: NonZeroUsize) -> bool

impl<'a, B: BitBlock + 'a> BitSlice<B>

pub fn bit_or_assign_raw<I: IntoIterator<Item = &'a B>>(&mut self, other: I)

Does a BitOrAssign between rhs and self. If rhs is bigger than self, the overflowing bits are ignored.
This function is purely a convinience function for Self::bit_or_assign_iter.

Example

let mut foo = BitMap::from([0b00110000u8, 0b00010010u8]);

foo.bit_or_assign_raw(&[0b10000000]);
assert_eq!(foo.test(0), Some(true));
assert_eq!(foo.test(8), Some(false));

impl<B: BitBlock> BitSlice<B>

pub unsafe fn test_unchecked(&self, idx: usize) -> bool

Returns true if the bit at idx is 1.

Safety

This function is safe if and only if idx < self.len_blocks().

pub unsafe fn set_unchecked(&mut self, idx: usize)

Sets the bit at idx to 1.

Safety

This function is safe if and only if idx < self.len_blocks().

pub unsafe fn reset_unchecked(&mut self, idx: usize)

Sets the bit at idx to 0

Safety

This function is safe if and only if idx < self.len_blocks().

pub unsafe fn fill_ones_unchecked(&mut self, start: usize, length: NonZeroUsize)

Fills a region starting at bit index start with 1s. TODO: test

Safety

TODO

pub unsafe fn fill_zeros_unchecked( &mut self, start: usize, length: NonZeroUsize )

Fills a region starting at bit index start with 0s. TODO: test

Safety

TODO